requirements.txt
git+https://github.com/tulir/mautrix-appservice-python.git
sqlalchemy
PyYAML